refactor: unify SIDE_EFFECT_TOOLS into single source of truth
